fix: address PR review — localization, redaction, config-sourced padding, test-seam removal, geometry refactor, selection extraction, test coverage
- Remove `simulateHardwarePressForTesting` from Sources; tests drive
  pressesBegan/pressesEnded directly via @testable import.
- Localize the two new strings (en+ja) in Localizable.xcstrings; copy toast
  uses an ICU plural (.one/.other).
- Redact TerminalInputDebugLog: drop textSummary, log byte counts only, never
  typed content/bytes (keyCode/mods/handled/len only).
- Source Ghostty window padding from a single owner (GhosttyRuntime) that both
  generates the config and feeds the selection inset; drop duplicated constants
  (ghostty_config_get cannot vend WindowPadding).
- Replace caseless-enum TerminalSelectionGeometry with methods on
  TerminalGridCell / TerminalSelectionCellGeometry.
- Extract the drag-select/copy/HUD subsystem into GhosttySurfaceView+Selection.swift.
- Balance pressesBegan/Ended/Cancelled forwarding via a captured-press set.
- Add tests: Option+Up/Down consumed-but-silent; pressesCancelled cancels repeat.
