+++ Packages/iOS/CmuxMobileTerminal/Sources/CmuxMobileTerminal/TerminalSelectionGeometry.swift
@@ -17,7 +17,7 @@ struct TerminalGridCell: Equatable, Sendable {
 
 /// Cell metrics in POINT space (already divided by the screen scale), describing
 /// how a viewport cell maps to a rect in the host view's coordinate system. The
-/// mapping is the exact inverse of ``TerminalSelectionGeometry/cell(at:geometry:)``
+/// mapping is the exact inverse of ``TerminalSelectionCellGeometry/cell(at:)``
 /// (the math `GhosttySurfaceView.scrollCell(at:)` runs): a cell `(col, row)`
 /// occupies the half-open box
 /// `[origin.x + col·cellWidth, origin.x + (col+1)·cellWidth)` ×
@@ -52,48 +52,48 @@ struct TerminalSelectionCellGeometry: Equatable, Sendable {
     }
 }
 
-/// Pure selection math shared by the drag-selection overlay and the local
-/// text-extraction read. Deliberately UIKit-free (CoreGraphics only) so the
-/// cell↔rect and anchor/focus↔ordered-range logic is unit-testable without a
-/// surface, a window, or a simulator.
+extension TerminalGridCell {
+    /// True when `self` is at or before `other` in row-major reading order.
+    func precedes(orEqualTo other: TerminalGridCell) -> Bool {
+        row != other.row ? row < other.row : col <= other.col
+    }
+
+    /// This cell clamped into `[0, columns-1] × [0, rows-1]`. A drag can run past
+    /// the last column or below the last row (the recognizer keeps reporting
+    /// points outside the render rect), so every cell is clamped before it drives
+    /// a rect or a text range. `columns`/`rows` of 0 collapse to the origin cell.
+    func clamped(toColumns columns: Int, rows: Int) -> TerminalGridCell {
+        let maxCol = max(columns - 1, 0)
+        let maxRow = max(rows - 1, 0)
+        return TerminalGridCell(
+            col: min(max(col, 0), maxCol),
+            row: min(max(row, 0), maxRow)
+        )
+    }
+}
+
+/// Pure selection math hung off the cell geometry that owns the metrics, shared
+/// by the drag-selection overlay and the local text-extraction read. Deliberately
+/// UIKit-free (CoreGraphics only) so the cell↔rect and anchor/focus↔ordered-range
+/// logic is unit-testable without a surface, a window, or a simulator.
 ///
 /// The iPad terminal is a thin mirror of the Mac's ghostty surface, so the
 /// on-device ghostty selection is never visible (the Mac owns it and re-streams
 /// highlighted cells over the render-grid). The drag selection is therefore
-/// rendered and tracked entirely from these functions; see
+/// rendered and tracked entirely from these methods; see
 /// ``GhosttySurfaceView/handleSelectionPan(_:)``.
-enum TerminalSelectionGeometry {
-    /// True when `lhs` is at or before `rhs` in row-major reading order.
-    static func precedesOrEqual(_ lhs: TerminalGridCell, _ rhs: TerminalGridCell) -> Bool {
-        lhs.row != rhs.row ? lhs.row < rhs.row : lhs.col <= rhs.col
-    }
-
-    /// Clamp a cell into `[0, columns-1] × [0, rows-1]`. A drag can run past the
-    /// last column or below the last row (the recognizer keeps reporting points
-    /// outside the render rect), so every cell is clamped before it drives a
-    /// rect or a text range. `columns`/`rows` of 0 collapse to the origin cell.
-    static func clamp(_ cell: TerminalGridCell, columns: Int, rows: Int) -> TerminalGridCell {
-        let maxCol = max(columns - 1, 0)
-        let maxRow = max(rows - 1, 0)
-        return TerminalGridCell(
-            col: min(max(cell.col, 0), maxCol),
-            row: min(max(cell.row, 0), maxRow)
-        )
-    }
-
+extension TerminalSelectionCellGeometry {
     /// Order an `anchor` (drag start) and `focus` (current/end) pair into the
-    /// inclusive `(start, end)` range in reading order, clamping both to the
-    /// grid. `start` is always at or before `end`, so it maps directly to a
-    /// ghostty selection's `top_left` / `bottom_right`.
-    static func normalizedRange(
+    /// inclusive `(start, end)` range in reading order, clamping both to the grid.
+    /// `start` is always at or before `end`, so it maps directly to a ghostty
+    /// selection's `top_left` / `bottom_right`.
+    func normalizedRange(
         anchor: TerminalGridCell,
-        focus: TerminalGridCell,
-        columns: Int,
-        rows: Int
+        focus: TerminalGridCell
     ) -> (start: TerminalGridCell, end: TerminalGridCell) {
-        let a = clamp(anchor, columns: columns, rows: rows)
-        let b = clamp(focus, columns: columns, rows: rows)
-        return precedesOrEqual(a, b) ? (a, b) : (b, a)
+        let a = anchor.clamped(toColumns: columns, rows: rows)
+        let b = focus.clamped(toColumns: columns, rows: rows)
+        return a.precedes(orEqualTo: b) ? (a, b) : (b, a)
     }
 
     /// The highlight rectangles (one per visual row) for the inclusive selection
@@ -104,27 +104,21 @@ enum TerminalSelectionGeometry {
     /// - middle rows: `0 … lastColumn` (full width)
     /// - last row:   `0 … end.col`
     ///
-    /// Rects are in the host view's point space, computed from `geometry`.
-    static func selectionRects(
+    /// Rects are in the host view's point space, computed from this geometry.
+    func selectionRects(
         anchor: TerminalGridCell,
-        focus: TerminalGridCell,
-        geometry: TerminalSelectionCellGeometry
+        focus: TerminalGridCell
     ) -> [CGRect] {
-        let (start, end) = normalizedRange(
-            anchor: anchor,
-            focus: focus,
-            columns: geometry.columns,
-            rows: geometry.rows
-        )
-        let lastColumn = max(geometry.columns - 1, 0)
+        let (start, end) = normalizedRange(anchor: anchor, focus: focus)
+        let lastColumn = max(columns - 1, 0)
 
         func rowRect(row: Int, fromCol: Int, toCol: Int) -> CGRect {
             let lo = min(fromCol, toCol)
             let hi = max(fromCol, toCol)
-            let x = geometry.origin.x + CGFloat(lo) * geometry.cellWidth
-            let y = geometry.origin.y + CGFloat(row) * geometry.cellHeight
-            let width = CGFloat(hi - lo + 1) * geometry.cellWidth
-            return CGRect(x: x, y: y, width: width, height: geometry.cellHeight)
+            let x = origin.x + CGFloat(lo) * cellWidth
+            let y = origin.y + CGFloat(row) * cellHeight
+            let width = CGFloat(hi - lo + 1) * cellWidth
+            return CGRect(x: x, y: y, width: width, height: cellHeight)
         }
 
         if start.row == end.row {
@@ -143,19 +137,17 @@ enum TerminalSelectionGeometry {
     }
 
     /// The viewport cell containing `point` (host-view point space) — the exact
-    /// inverse of ``selectionRects``' per-cell rect: `col = ⌊(x − origin.x) /
-    /// cellWidth⌋`, `row = ⌊(y − origin.y) / cellHeight⌋`, each clamped to ≥ 0
-    /// (the recognizer can report points left of / above the grid; the upper
-    /// bound is applied later by ``normalizedRange`` so a drag can run past the
-    /// last column/row). This is the math ``GhosttySurfaceView/scrollCell(at:)``
-    /// runs, kept here so the point→cell hit-test and the cell→rect highlight are
-    /// derived from ONE definition and cannot drift apart.
-    static func cell(
-        at point: CGPoint,
-        geometry: TerminalSelectionCellGeometry
-    ) -> TerminalGridCell {
-        let col = max(0, Int((point.x - geometry.origin.x) / geometry.cellWidth))
-        let row = max(0, Int((point.y - geometry.origin.y) / geometry.cellHeight))
+    /// inverse of ``selectionRects(anchor:focus:)``' per-cell rect: `col = ⌊(x −
+    /// origin.x) / cellWidth⌋`, `row = ⌊(y − origin.y) / cellHeight⌋`, each clamped
+    /// to ≥ 0 (the recognizer can report points left of / above the grid; the upper
+    /// bound is applied later by ``normalizedRange(anchor:focus:)`` so a drag can
+    /// run past the last column/row). This is the math
+    /// ``GhosttySurfaceView/scrollCell(at:)`` runs, kept here so the point→cell
+    /// hit-test and the cell→rect highlight are derived from ONE definition and
+    /// cannot drift apart.
+    func cell(at point: CGPoint) -> TerminalGridCell {
+        let col = max(0, Int((point.x - origin.x) / cellWidth))
+        let row = max(0, Int((point.y - origin.y) / cellHeight))
         return TerminalGridCell(col: col, row: row)
     }
 }
